CLEM: changed event status info from integer to double

+++ src/main/java/at/naurandir/discord/clem/bot/service/client/dto/worldstate/EventDTO.java
@@ -26,6 +26,6 @@ public class EventDTO {
     
     private String victimNode;
     
-    private Integer health;
+    private Double health;
     
 }
